+++ zcash_transparent/CHANGELOG.md
@@ -13,6 +13,14 @@ workspace.
 ### Changed
 - MSRV is now 1.85.1.
 
+## [0.6.1] - 2025-10-27
+
+### Added
+- `zcash_transparent::address::TransparentAddress::to_zcash_address`
+- `zcash_transparent::pczt`:
+  - `Bip32Derivation::extract_zip_48_fields`
+- `zcash_transparent::zip48` module, behind the `transparent-inputs` feature flag.
+
 ## [0.6.0] - 2025-10-02
 
 ### Added